Migrate setting commands
Moved commands related to settings to a new cog named Settings.
